# dbBulkCopy feature enhancement