--- a/R/DBIConvenience.R
+++ b/R/DBIConvenience.R
@@ -225,7 +225,7 @@ bulk.copy <- function(con,name,value,...){
 bulk.copy.file <- function(con,name,value,headers=TRUE,delim=","){
   con.string = dbGetInfo(con)$ConnectionString
   if (!is.null(value) && file.exists(value))
-    clrCallStatic("rsqlserver.net.misc","SqlBulkCopy",con.string ,value,name,headers,delim)
+    lapply(value, function(x) clrCallStatic("rsqlserver.net.misc","SqlBulkCopy",con.string,x,name,headers,delim))
   else
     stop("one or more files are null or do not exist")