**Pull Request Changes**
- [x] Data frames containing NAs upload as NULL
- [x] Commas inside quoted text fields are not considered delimiters. Fixes #25 
- [x] Quotes are not uploaded even for quoted text fields. Fixes #18 
- [x] Support tab- and pipe-delimited text files
- [ ] Measure changes in performance