@ruaridhw 2016-12-17
@agstudy The current master will not function correctly as the [dll](https://github.com/agstudy/rsqlserver/blob/master/inst/libs/rsqlserver.net.dll) file is not representative of the new C# source. (I don't believe the Makefiles or install.libs.R are being called on installation as opposed to merely replicating inst/libs?)  I have just pushed the correct binaries and Lumenworks references to my fork ruaridhw/rsqlserver@06c35e1f0.